Audio file and placeholders for Zelda II
This is an older template I put together a few years ago (2014). I do
not know if OGG looping is supported or not in the latest HDNes but if
not intro songs will play over again from the begining which is not
correct behavior. Anyways this one was super complicated to compile and
configure so hopefully it helps who ever uses it! Enjoy.